backend/gateway/requirements.txt
# GraphQL Gateway 의존성

# FastAPI 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# HTTP 클라이언트
aiohttp==3.9.1
httpx==0.25.2

# 유틸리티
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 로깅
structlog==23.2.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 개발 도구
black==23.11.0
flake8==6.1.0
mypy==1.7.1
